+++ lib/types.ts
@@ -4,7 +4,7 @@ import DomElement from './models/dom_element';
 export type CssSelector = string;
 export type XPathSelector = string;
 
-export type WendigoSelector = CssSelector | XPathSelector | DomElement;
+export type WendigoSelector = string | DomElement;
 
 export interface BrowserSettings {
     log?: boolean;