+++ tests/browser/find_by_text.test.js
@@ -3,6 +3,7 @@
 const assert = require('assert');
 const Wendigo = require('../..');
 const configUrls = require('../config.json').urls;
+const utils = require('../test_utils');
 
 describe("Find By Text", function() {
     this.timeout(5000);
@@ -38,7 +39,7 @@ describe("Find By Text", function() {
     });
 
     it("Find By Text SubQuery With Nested Elements", async() => {
-        const titleElements = await browser.findByTextContaining("body", "My first paragraph");
+        const titleElements = await browser.findByText("body", "My first paragraph");
         assert.strictEqual(titleElements.length, 1);
     });
 
@@ -48,10 +49,15 @@ describe("Find By Text", function() {
     });
 
     it("Find By Text Containing SubQuery", async() => {
-        const paragraphElements = await browser.findByTextContaining(".container", "first paragraph");
+        const paragraphElements = await browser.findByTextContaining(".container", "paragraph");
         assert.strictEqual(paragraphElements.length, 1);
         const titleElements = await browser.findByTextContaining(".container", "Title");
         assert.strictEqual(titleElements.length, 0);
+    });
+
+    it("Find By Text Containing SubQuery With Nested Elements", async() => {
+        const titleElements = await browser.findByTextContaining("body", "paragraph");
+        assert.strictEqual(titleElements.length, 2);
     });
 
     it("Find By Text Containing SubQuery With Nested Elements", async() => {
@@ -67,5 +73,53 @@ describe("Find By Text", function() {
         assert.strictEqual(clicked, 3);
     });
 
-    it("Find By Text With Invalid Selector");
+    it("Find By Text With Invalid Selector", async() => {
+        await utils.assertThrowsAsync(async() => {
+            await browser.findByText({}, "first");
+        }, "Error: [findByText] Invalid selector.");
+    });
+
+    it("Find By Text Containing With Invalid Selector", async() => {
+        await utils.assertThrowsAsync(async() => {
+            await browser.findByTextContaining({}, "first");
+        }, "Error: [findByTextContaining] Invalid selector.");
+    });
+
+    // Tests #342
+    it.skip("Find By Text With Quotes Character", async() => {
+        await browser.open(configUrls.weirdText);
+        const elems = await browser.findByText("\"quotes'");
+        assert.strictEqual(elems.length, 1);
+    });
+
+    it("Find By Text With Quotes Character Fails", async() => {
+        await browser.open(configUrls.weirdText);
+        const elems = await browser.findByText("\"quo");
+        assert.strictEqual(elems.length, 0);
+    });
+
+    it("Find By Text Containing With Quotes Character", async() => {
+        await browser.open(configUrls.weirdText);
+        const elems = await browser.findByTextContaining("\"quo");
+        assert.strictEqual(elems.length, 1);
+    });
+
+    // Tests #342
+    it.skip("Find By Text Containing With Simple Quote Character", async() => {
+        await browser.open(configUrls.weirdText);
+        const elems = await browser.findByTextContaining("es'");
+        assert.strictEqual(elems.length, 1);
+    });
+
+    it("Find By Text With Colon Character", async() => {
+        await browser.open(configUrls.weirdText);
+        const elems = await browser.findByText("colon:");
+        assert.strictEqual(elems.length, 1);
+    });
+
+    it("Find By Text Containing With Colon Character", async() => {
+        await browser.open(configUrls.weirdText);
+        const elems = await browser.findByTextContaining("on:");
+        assert.strictEqual(elems.length, 1);
+    });
 });