+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Wendigo refactored into TypeScript, support for types files
 * Browser.query and browser.queryAll now support XPath selectors
 * Browser.queryXPath removed
+* Minor fixes to xPath queries
 * DomElement class methods changed
 * Complex subqueries with multiple parent elements supported
 * New methods added to DomElement and subqueries improved